add endpoint update_attendance, delete_attendance

+++ dawis/src/main/java/com/nawasenaproject/dawis/dto/AttendanceResponse.java
@@ -19,7 +19,7 @@ public class AttendanceResponse {
 
     private Integer specialWage;
 
-    private Integer mandays;
+    private Double mandays;
 
     private Integer bonuses;
